pyproject.toml
[tool.poetry]
name = "rf-loaddotenv"
version = "0.1.0"
description = "Demonstração de carregamento de arquivos .env"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
robotframework = "^7.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
